renamed stix_to_<module_name> to stix_to_query to keep things consistent between adapters
