+++ stix_shifter/stix_translation/src/modules/bigfix/bigfix_translator.py
@@ -1,8 +1,9 @@
 from ..base.base_translator import BaseTranslator
-from .stix_to_bigfix import StixToRelevanceQuery
+from .stix_to_query import StixToQuery
 from ...json_to_stix.json_to_stix import JSONToStix
 
 from os import path
+
 
 class Translator(BaseTranslator):
 
@@ -10,7 +11,7 @@ class Translator(BaseTranslator):
         basepath = path.dirname(__file__)
         filepath = path.abspath(
             path.join(basepath, "json", "to_stix_map.json"))
-        
+
         self.mapping_filepath = filepath
         self.result_translator = JSONToStix(filepath)
-        self.query_translator = StixToRelevanceQuery()
+        self.query_translator = StixToQuery()
