+++ stix_shifter/stix_translation/src/modules/csa/csa_translator.py
@@ -1,9 +1,10 @@
 from ..base.base_translator import BaseTranslator
-from .stix_to_cloudsql import StixToCloudSQL
+from .stix_to_query import StixToQuery
 from ...json_to_stix.json_to_stix import JSONToStix
 
 
 from os import path
+
 
 class Translator(BaseTranslator):
     def __init__(self, dialect=None):
@@ -12,4 +13,4 @@ class Translator(BaseTranslator):
             path.join(basepath, "json", "to_stix_map.json"))
         self.mapping_filepath = filepath
         self.result_translator = JSONToStix(filepath)
-        self.query_translator = StixToCloudSQL(dialect=dialect)
+        self.query_translator = StixToQuery(dialect=dialect)
