+++ stix_shifter/stix_translation/src/modules/qradar/qradar_translator.py
@@ -1,7 +1,7 @@
 
 # ..base.base_translator import BaseTranslator
 from ..base.base_translator import BaseTranslator
-from .stix_to_aql import StixToAQL
+from .stix_to_query import StixToQuery
 from ...json_to_stix.json_to_stix import JSONToStix
 from .qradar_utils import hash_type_lookup
 
@@ -17,4 +17,4 @@ class Translator(BaseTranslator):
         self.mapping_filepath = filepath
         # Pass in callback function to handle hashes with unknown type
         self.result_translator = JSONToStix(filepath, hash_type_lookup)
-        self.query_translator = StixToAQL()
+        self.query_translator = StixToQuery()
