+++ stix_shifter/stix_translation/src/modules/splunk/splunk_translator.py
@@ -1,7 +1,7 @@
 # ..base.base_translator import BaseTranslator
 from ..base.base_translator import BaseTranslator
 from ...json_to_stix.json_to_stix import JSONToStix
-from .stix_to_splunk import StixToSplunk
+from .stix_to_query import StixToQuery
 from .splunk_utils import hash_type_lookup
 
 from os import path
@@ -15,4 +15,4 @@ class Translator(BaseTranslator):
             path.join(basepath, "json", "to_stix_map.json"))
         self.mapping_filepath = filepath
         self.result_translator = JSONToStix(filepath, hash_type_lookup)
-        self.query_translator = StixToSplunk()
+        self.query_translator = StixToQuery()
